# [CANN] Fix Multi-NPU execution error